% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_by.R
\name{multiply_by}
\alias{multiply_by}
\title{Print n*multiplier}
\usage{
multiply_by(n, multiplier = c("two", "three", "four"))
}
\arguments{
\item{n}{A numeric definign the number that should be multiplied}

\item{multiplier}{either two, three or four}
}
\value{
This function returns n*multiplier
}
\description{
This is a simple function that multiplies a number by multiplier. You can
change the number input (using the \code{n} argument) and choose
a suitable multiolier (\code{multiplier}).
}
\examples{
multiply_by(5, "two")
multiply_by(5, "three")
multiply_by(5, "four")

}
